changed endpoints name to follow best practices

+++ training2/src/main/java/bbva/training2/repository/BookRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 
 public interface BookRepository extends JpaRepository<Book, Long> {
 
-    Optional<Book> findByIdBook(int id);
+    Optional<Book> findByIdBook(long id);
 
     Book findByTitle(String title);
